+++ src/types/tag.rs
@@ -3,5 +3,6 @@ use crate::types::HeadingLevel;
 #[derive(Debug, PartialEq)]
 pub enum Tag {
     Heading(HeadingLevel),
+    IndentedCodeBlock,
     Paragraph,
 }